Command line: refactored subcommands into separate classes.
